+++ src/addons/immortal_msk.hpp
@@ -10,7 +10,7 @@
 
 namespace addon {
     class immortal_class {
-        typedef size_t index_type;
+        typedef unsigned index_type;
     public:
         //------------------- ctors -------------------
         immortal_class(): file_switch_(2) //"invalid" at start
